Backend/requirements.txt
asgiref
Django
django-cors-headers
django-environ
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
tzdata
python-dotenv
Pillow

# Successfully installed Django-5.0.4 PyJWT-2.8.0 asgiref-3.8.1 django-cors-headers-4.3.1 djangorestframework-3.15.1 djangorestframework-simplejwt-5.3.1 psycopg2-binary-2.9.9 python-dotenv-1.0.1 pytz-2024.1 sqlparse-0.5.0 typing-extensions-4.11.0

